+++ src/argilla_sdk/records/_dataset_records.py
@@ -307,7 +307,7 @@ class DatasetRecords(Iterable[Record], LoggingMixin):
         """
         return self(with_suggestions=True, with_responses=True).to_list(flatten=flatten)
 
-    def to_json(self, path: Union[Path, str]) -> str:
+    def to_json(self, path: Union[Path, str]) -> Path:
         """
         Export the records to a file on disk. This is a convenient shortcut for dataset.records(...).to_disk().
 
@@ -319,16 +319,20 @@ class DatasetRecords(Iterable[Record], LoggingMixin):
             The path to the file where the records were saved.
 
         """
+        if isinstance(path, str):
+            path = Path(path)
+        if path.exists():
+            raise FileExistsError(f"File {path} already exists.")
         record_dicts = self(with_suggestions=True, with_responses=True).to_list()
         with open(path, "w") as f:
             json.dump(record_dicts, f)
         return path
 
-    def from_json(self, path: str) -> "DatasetRecords":
+    def from_json(self, path: Union[Path, str]) -> "DatasetRecords":
         """Creates a DatasetRecords object from a disk path.
 
         Args:
-            path (str): The path to the dataset records.
+            path (str): The path to the file containing the records.
 
         Returns:
             DatasetRecords: The DatasetRecords object created from the disk path.
